+++ tdvt/test/tdvt_test.py
@@ -228,6 +228,11 @@ class ReRunFailedTestsTest(BaseTDVTTest):
         all_test_results = tdvt_core.run_tests_serial(tests)
         self.check_results(all_test_results, 5)
 
+    def test_failed_rerun(self):
+        tests = enqueue_failed_tests(Path(get_path('tool_test/rerun_failed_tests', 'failed_tests.json', __name__)),
+                                     TEST_DIRECTORY, None, self.test_config.tested_run_time_config)
+        self.assertTrue(tests[0][0].get_expected_message() == "Invalid username or password", "Expected message was not read in correctly.")
+
     def test_logical_rerun_fail(self):
         tests = enqueue_failed_tests(Path(get_path('tool_test/rerun_failed_tests', 'logical_compare_sql.json', __name__)),
                                      TEST_DIRECTORY, None, self.test_config.tested_run_time_config)
@@ -864,7 +869,37 @@ class ResultsExceptionTest(unittest.TestCase):
             actual_message = mock_batch.results[test_file].get_failure_message_or_all_exceptions()
             self.assertTrue(isinstance(mock_batch.results[test_file].test_case_map[0].error_type, error_state))
             self.assertTrue(mock_batch.results[test_file].test_case_map[0].all_passed() == True)
+    def test_process_error_output_json_expected(self):
+        error_state = TestErrorExpected
+        proc_error_code = 1
+        test_set_expected = MockTestSet('not_used', 'not used', 'mock ds', 'tests', 'mock config', 'mock.tds', '', 'tests/*.txt', False, 'mock suite expression', '', '')
+        test_set_expected.expected_message = 'Invalid username or password'
+        error_message = test_set_expected.expected_message
+        mock_batch = MockBatchQueueWork(self.mock_tests, self.test_config, test_set_expected, subprocess.CalledProcessError(proc_error_code, 'test', error_message))
+        self.check_errors(error_message, error_state, mock_batch)
 
+        for test_file in mock_batch.results:
+            json_str = json.dumps(mock_batch.results[test_file], cls=TestOutputJSONEncoder)
+            json_object = json.loads(json_str)
+            self.assertEqual(json_object['expected_message'], error_message)
+            self.assertTrue(isinstance(mock_batch.results[test_file].test_case_map[0].error_type, error_state))
+            self.assertTrue(mock_batch.results[test_file].test_case_map[0].all_passed() == True)
+
+    def test_process_error_output_json_not_expected(self):
+            error_state = TestErrorExpected
+            proc_error_code = 1
+            test_set_expected = MockTestSet('not_used', 'not used', 'mock ds', 'tests', 'mock config', 'mock.tds', '', 'tests/*.txt', False, 'mock suite expression', '', '')
+            test_set_expected.expected_message = 'Invalid username or password'
+            error_message = test_set_expected.expected_message
+            mock_batch = MockBatchQueueWork(self.mock_tests, self.test_config, test_set_expected, subprocess.CalledProcessError(proc_error_code, 'test', error_message))
+            self.check_errors(error_message, error_state, mock_batch)
+
+            for test_file in mock_batch.results:
+                json_str = json.dumps(mock_batch.results[test_file], cls=TestOutputJSONEncoder)
+                json_object = json.loads(json_str)
+                self.assertEqual(json_object['expected_message'], error_message)
+                self.assertTrue(isinstance(mock_batch.results[test_file].test_case_map[0].error_type, error_state))
+                self.assertTrue(mock_batch.results[test_file].test_case_map[0].all_passed() == True)
 
     def test_process_timeout_multiple(self):
         test_name = 'setup.mytest.txt'